Patterns as a separate project Contributing.md added Readme.md update
