+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+# Contributing
+When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owners of this repository before making a change.
+## Setting Up the project locally
+## Submitting a Pull Request
+* Please go through existing issues and pull requests to check if somebody else is already working on it.
+* Increase the version numbers.
+* You may merge the Pull Request in once you have the sign-off of other developer, or if you do not have permission to do that, you may request to the reviewer to merge it for you.
+## Code of conduct
+This project has adopted the [Microsoft Open Source Code of Conduct](https://opensource.microsoft.com/codeofconduct/).
