+++ Jal.Router.Azure.Standard.LightInject.Installer.All/Jal.Router.Azure.Standard.LightInject.Installer.All.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <Version>2.0.1</Version>
+    <Version>2.1.0</Version>
     <Authors>Raul Naupari</Authors>
     <Company>Raul Naupari</Company>
     <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
@@ -22,6 +22,7 @@
     <ProjectReference Include="..\Jal.Router.Logger.LightInject.Installer\Jal.Router.Logger.LightInject.Installer.csproj" />
     <ProjectReference Include="..\Jal.Router.Logger\Jal.Router.Logger.csproj" />
     <ProjectReference Include="..\Jal.Router.Newtonsoft.LightInject.Installer\Jal.Router.Newtonsoft.LightInject.Installer.csproj" />
+    <ProjectReference Include="..\Jal.Router.Patterns\Jal.Router.Patterns.csproj" />
     <ProjectReference Include="..\Jal.Router\Jal.Router.csproj" />
   </ItemGroup>
