+++ Jal.Router.Patterns/Interface/IMessageHandlerWithData.cs
@@ -1,7 +1,7 @@
 using Jal.Router.Model;
 using System.Threading.Tasks;
 
-namespace Jal.Router.Interface.Patterns
+namespace Jal.Router.Patterns.Interface
 {
     public interface IMessageHandlerWithData<in TMessage, in TData> : IMessageHandler<TMessage>
     {
