+++ Jal.Router/Jal.Router.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0</TargetFrameworks>
     <Authors>Raul Naupari</Authors>
-    <Version>2.0.0</Version>
+    <Version>2.1.0</Version>
     <Company>Raul Naupari</Company>
     <Description>Just another library to route messages</Description>
     <Copyright />
